no longer needs the mixin, just a special helper
